requirements.txt
absl-py==0.9.0
ase==3.19.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
future==0.18.2
gitdb==4.0.5
gitpython==3.1.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
googledrivedownloader==0.4
gql==0.2.0
graphql-core==1.1
grpcio==1.30.0
h5py==2.10.0
hydra-core==0.11.3
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0; python_version < "3.8"
isodate==0.6.0
jinja2==2.11.2
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
kiwisolver==1.2.0
llvmlite==0.32.1
markdown==3.2.2
markupsafe==1.1.1
matplotlib==3.2.2
networkx==2.4
numba==0.49.1
numpy==1.19.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
omegaconf==1.4.1
pandas==1.0.5
pathtools==0.1.2
pillow==7.2.0
plyfile==0.7.2
promise==2.3
protobuf==3.12.2
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-metric-learning==0.9.89.dev1
pytz==2020.1
pywavelets==1.1.1
pyyaml==5.3.1
pyzmq==19.0.1
rdflib==5.0.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6; python_version >= "3"
scikit-image==0.16.2
scikit-learn==0.23.1
scipy==1.5.1
sentry-sdk==0.16.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
torch==1.5.1
torch-cluster==1.5.5
torch-geometric==1.6.0
torch-points-kernels==0.6.6
torch-scatter==2.0.5
torch-sparse==0.6.6
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.6.1
tornado==6.0.4
tqdm==4.47.0
urllib3==1.25.9
visdom==0.1.8.9
wandb==0.8.36
watchdog==0.10.3
websocket-client==0.57.0
werkzeug==1.0.1
wheel==0.34.2; python_version >= "3"
zipp==3.1.0; python_version < "3.8"
